requirements.txt
tkinterdnd2
Pillow
PyMuPDF
PyPDF2
python-docx
watchdog
extract_msg
pywin32